requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
annotated-types
anyio
attrs
audioread
certifi
cffi
charset-normalizer
click
colorama
cryptography
deprecation
distro
fastapi
frozenlist
h11
h2
hpack
httpcore
httpx
hyperframe
idna
jiter
loguru
multidict
numpy
openai
packaging
postgrest
propcache
pyactiveresource
PyAudio
pycparser
pydantic
pydantic_core
pydub
PyJWT
python-dotenv
python-multipart
PyYAML
realtime
requests
scipy
ShopifyAPI
six
sniffio
starlette
storage3
StrEnum
supabase
supabase-auth
supabase-functions
tqdm
twilio
typing_extensions
typing-inspection
urllib3
uvicorn
websockets
win32_setctime
yarl
uvloop
gunicorn
aiofiles
python-json-logger
tenacity
jinja2
psycopg2-binary
elevenlabs


sounddevice
soundfile
